+++ packages/terra-framework-docs/src/terra-dev-site/doc/table/Examples/TableWithoutHeaders.jsx
@@ -56,7 +56,7 @@ const TableWithoutHeaders = () => {
       id="table-without-headers"
       overflowColumns={cols}
       rows={rows}
-      hasColumnHeaders={false}
+      hasVisibleColumnHeaders={false}
     />
   );
 };