+++ packages/terra-table/src/subcomponents/ColumnHeader.jsx
@@ -59,11 +59,11 @@ const propTypes = {
   /**
    * Boolean indicating whether or not the table columns should be displayed.
    */
-  hasColumnHeaders: PropTypes.bool,
+  hasVisibleColumnHeaders: PropTypes.bool,
 };
 
 const defaultProps = {
-  hasColumnHeaders: true,
+  hasVisibleColumnHeaders: true,
 };
 
 const ColumnHeader = (props) => {
@@ -78,15 +78,14 @@ const ColumnHeader = (props) => {
     onColumnSelect,
     onResizeMouseDown,
     onResizeHandleChange,
-    hasColumnHeaders,
+    hasVisibleColumnHeaders,
   } = props;
 
   return (
     <thead>
       <tr
-        aria-rowindex={1}
-        className={cx('column-header-row', { hidden: !hasColumnHeaders })}
-        height={hasColumnHeaders ? headerHeight : undefined}
+        className={cx('column-header-row', { hidden: !hasVisibleColumnHeaders })}
+        height={hasVisibleColumnHeaders ? headerHeight : undefined}
       >
         {columns.map((column, columnIndex) => (
           <ColumnHeaderCell
@@ -101,8 +100,8 @@ const ColumnHeader = (props) => {
             minimumWidth={column.minimumWidth}
             maximumWidth={column.maximumWidth}
             headerHeight={headerHeight}
-            isResizable={hasColumnHeaders && column.isResizable}
-            isSelectable={hasColumnHeaders && column.isSelectable}
+            isResizable={hasVisibleColumnHeaders && column.isResizable}
+            isSelectable={hasVisibleColumnHeaders && column.isSelectable}
             tableHeight={tableHeight}
             isActive={activeColumnIndex === columnIndex}
             isResizeActive={activeColumnIndex === columnIndex && isActiveColumnResizing}