+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,9 +4,12 @@
 
 * Added
   * Added example to `terra-table` to show column states and sorting;
-  
+  * Added `hasVisibleColumnHeaders` example for FlowsheetDataGrid.
+
 * Updated
   * Updated About page of `terra-table` to provide accessibility documentation for Home and End keys.
+  * Removed alpha notice for FlowsheetDataGrid.
+
 
 ## 1.44.0 - (November 9, 2023)